refactor(spec): route spec-decode emitter + target through Carrier; move emitters to arch crates
Completes the dependency-inversion of the SpecTarget seam (V3b + V6): the
daemon no longer arch-matches for the spec-decode path, and the per-arch
emitter definitions live in their arch crates.

- New `SpecEmitCtx` (hipfire-runtime::spec): the model-independent recipe
  generate_spec builds after slot acquisition (tokenizer + slot eos + raw
  tool JSON + reasoning level + pre-decoded vocab). No arch type appears.

- Qwen35Emit → hipfire-arch-qwen35::spec_emit, Deepseek4Emit →
  hipfire-arch-deepseek4::spec_emit. Each gets `from_ctx(SpecEmitCtx) ->
  Box<dyn SpecEmit>` that extracts its own grammar ToolSchema from the raw
  tool JSON internally (that JSON→schema build moved out of the daemon too).
  Qwen35Emit is the shared ChatML emitter — the llama/qwen2 n-gram arms drive
  it too (they already share qwen35's tool-call grammar).

- Carrier (hipfire-loader) gains `spec_target_guard` + `make_spec_emitter`,
  plus `carrier_for(arch_id)`. generate_spec resolves the carrier once and
  routes BOTH the target borrow and the emitter through it; the EmitSpec enum
  and the daemon's two emitter definitions are deleted. The free-fn
  spec_target_guard arch_id match in spec_build.rs is removed.

build_speculator stays the loader-internal load-time registry (it is never a
daemon coupling — only carriers' load() call it); hoisting it onto Carrier is
a documented follow-up.

Gates (gfx1100): coherence 11/11 OK, serve-multiturn PASS (AR + DFlash 27B,
#462 guard intact), deepseek4-mtp OK (both rows, accept 86%/74%).
